refactor: restructure Motion
